fix: release note popup
