+++ packages/vscode-extension/src/extension.ts
@@ -61,8 +61,7 @@ import {
 } from "./utils/commonUtils";
 import { loadLocalizedStrings } from "./utils/localizeUtils";
 import { ExtensionSurvey } from "./utils/survey";
-import { ExtensionUpgrade } from "./utils/upgrade";
-import { PrereleasePage } from "./utils/prerelease";
+import { ReleaseNote } from "./utils/releaseNote";
 
 export let VS_CODE_UI: VsCodeUI;
 
@@ -888,10 +887,8 @@ async function runBackgroundAsyncTasks(
   ExtTelemetry.settingsVersion = await handlers.getSettingsVersion();
 
   await ExtTelemetry.sendCachedTelemetryEventsAsync();
-  const upgrade = new ExtensionUpgrade(context);
-  await upgrade.showChangeLog();
-  const prereleasePage = new PrereleasePage(context);
-  await prereleasePage.checkAndShow();
+  const releaseNote = new ReleaseNote(context);
+  await releaseNote.show();
 
   await openWelcomePageAfterExtensionInstallation();
